Replace old widget-store files
