+++ widget-store-standalone/.devenv/profile
@@ -1,0 +1,1 @@
+/nix/store/96syj1xmrvw5x1n66mrc9bcwak0qphgx-devenv-profile
