--- widget-store-standalone/.eslintignore
@@ -1,4 +1,0 @@
-dist
-*.test.ts
-jest.config.js
-
