--- widget-store-standalone/.vscode/extensions.json
@@ -1,7 +1,0 @@
-{
-	// See http://go.microsoft.com/fwlink/?LinkId=827846
-	// for the documentation about the extensions.json format
-	"recommendations": [
-		"dbos-inc.dbos-ttdbg"
-	]
-}
