--- widget-store-standalone/.vscode/tasks.json
@@ -1,17 +1,0 @@
-{
-  "version": "2.0.0",
-  "tasks": [
-    {
-      "label": "npm: build",
-      "type": "npm",
-      "script": "build",
-      "group": {
-        "kind": "build",
-        "isDefault": true
-      },
-      "problemMatcher": [
-        "$tsc"
-      ]
-    }
-  ]
-}
