--- widget-store-standalone/dbos-config.yaml
@@ -1,29 +1,0 @@
-# To enable auto-completion and validation for this file in VSCode, install the RedHat YAML extension
-# https://marketplace.visualstudio.com/items?itemName=redhat.vscode-yaml
-
-# yaml-language-server: $schema=https://raw.githubusercontent.com/dbos-inc/dbos-transact/main/dbos-config.schema.json
-
-database:
-  hostname: "localhost"
-  port: 5432
-  username: "postgres"
-  password: ${PGPASSWORD}
-  app_db_name: "widget_store_standalone"
-  connectionTimeoutMillis: 3000
-  app_db_client: "knex"
-  migrate: ["npx knex migrate:latest", "npx knex seed:run"]
-  rollback: ["npx knex migrate:rollback"]
-telemetry:
-  logs:
-    logLevel: "info"
-application:
-  ses_from_address: ${REPORT_EMAIL_FROM_ADDRESS}
-  ses_to_address: ${REPORT_EMAIL_TO_ADDRESS}
-  aws_config:
-    aws_region: ${AWS_REGION}
-    aws_access_key_id: ${AWS_ACCESS_KEY_ID}
-    aws_secret_access_key: ${AWS_SECRET_ACCESS_KEY}
-env:
-  # If you want email reports nightly, set these
-  REPORT_EMAIL_FROM_ADDRESS: ${REPORT_EMAIL_FROM_ADDRESS}
-  REPORT_EMAIL_TO_ADDRESS: ${REPORT_EMAIL_TO_ADDRESS}
