--- widget-store-standalone/nodemon.json
@@ -1,7 +1,0 @@
-{
-  "watch": ["src/","migrations/","public/"],
-  "ext": "ts,json,liquid,js",
-  "ignore": ["src/**/*.test.ts"],
-  "exec": "npm run build && npx dbos migrate && npm run start"
-}
-
