--- widget-store/public/crash.liquid
@@ -1,107 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>DBOS Self-Destruct</title>
-    <link rel="icon" type="image/x-icon" href="https://dbos-blog-posts.s3.us-west-1.amazonaws.com/live-demo/favicon.ico">
-    <style>
-        body, html {
-            margin: 0;
-            padding: 0;
-            height: 100%;
-            background-color: #f5f5f5;
-            font-family: Arial, sans-serif;
-        }
-
-        .container {
-            display: flex;
-            flex-direction: column;
-            align-items: center;
-            padding: 20px;
-        }
-
-        .logo {
-            width: 200px;
-            height: auto;
-        }
-
-        .product {
-            background-color: #fff;
-            border: 1px solid #ddd;
-            border-radius: 8px;
-            padding: 20px;
-            margin: 20px 0;
-            width: 80%;
-            max-width: 600px;
-            text-align: center;
-        }
-
-        .product img {
-            max-width: 60%;
-            height: auto;
-            border-radius: 8px;
-        }
-
-        .product-title {
-            font-size: 24px;
-            color: #333;
-            margin: 20px 0;
-        }
-
-        .crash-button {
-            background-color: #f44336;
-            border: none;
-            color: white;
-            padding: 20px 40px; /* Increased padding */
-            text-align: center;
-            text-decoration: none;
-            display: inline-block;
-            font-size: 24px; /* Increased font size */
-            cursor: pointer;
-            border-radius: 8px;
-            transition: background-color 0.3s, box-shadow 0.3s;
-            margin-top: 20px;
-        }
-
-
-        .crash-button:hover {
-            background-color: #d32f2f;
-            box-shadow: 0 4px 8px 0 rgba(0,0,0,0.2);
-        }
-    </style>
-</head>
-<body>
-
-<div class="container">
-    <img src="https://dbos-blog-posts.s3.us-west-1.amazonaws.com/logos/black_logotype%2Btransparent_bg_h4000px.png" alt="DBOS Logo" class="logo">
-    <div class="product">
-        <img src="https://dbos-blog-posts.s3.us-west-1.amazonaws.com/live-demo/cartoon_bomb.png" alt="Cartoon Bomb">
-        <button class="crash-button" id="crashButton">Crash the Application</button>
-    </div>
-</div>
-
-<script>
-    document.getElementById('crashButton').addEventListener('click', async function() {
-        try {
-            const response = await fetch('/crash_application/', {
-                method: 'POST',
-                headers: {
-                    'Content-Type': 'application/json',
-                },
-            });
-
-            if (response.ok) {
-                const url = await response.text();
-            } else {
-                const errorText = await response.text();
-                console.error('Error:', errorText);
-            }
-        } catch (error) {
-            console.error('Error:', error);
-        }
-    });
-</script>
-
-</body>
-</html>
