--- widget-store/public/error.liquid
@@ -1,91 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Error - DBOS Widget Store</title>
-    <link rel="icon" type="image/x-icon" href="https://dbos-blog-posts.s3.us-west-1.amazonaws.com/live-demo/favicon.ico">
-    <style>
-        body, html {
-            margin: 0;
-            padding: 0;
-            height: 100%;
-            background-color: #f5f5f5;
-            font-family: Arial, sans-serif;
-        }
-
-        .container {
-            display: flex;
-            flex-direction: column;
-            align-items: center;
-            padding: 20px;
-        }
-
-        .logo {
-            width: 200px;
-            height: auto;
-        }
-
-        .confirmation-section {
-            background-color: #fff;
-            border: 1px solid #ddd;
-            border-radius: 8px;
-            padding: 20px;
-            margin: 20px 0;
-            width: 80%;
-            max-width: 600px;
-            text-align: center;
-        }
-
-        .section-title {
-            font-size: 24px;
-            color: #333;
-            margin-bottom: 30px;
-        }
-
-        .button-container {
-            display: flex;
-            justify-content: center;
-            gap: 20px;
-        }
-
-        .button {
-            background-color: #4CAF50;
-            border: none;
-            color: white;
-            padding: 25px 50px;
-            text-align: center;
-            text-decoration: none;
-            font-size: 22px;
-            cursor: pointer;
-            border-radius: 10px;
-            transition: background-color 0.3s, box-shadow 0.3s;
-        }
-
-        .button:hover {
-            box-shadow: 0 5px 15px 0 rgba(0,0,0,0.2);
-        }
-
-        .confirm-button {
-            background-color: #4CAF50;
-        }
-
-        .confirm-button:hover {
-            background-color: #45a049;
-        }
-    </style>
-</head>
-<body>
-
-<div class="container">
-    <img src="https://dbos-blog-posts.s3.us-west-1.amazonaws.com/logos/black_logotype%2Btransparent_bg_h4000px.png" alt="DBOS Logo" class="logo">
-    <div class="confirmation-section">
-        <div class="section-title">Error purchasing widget</div>
-        <div class="button-container">
-            <a href="/" class="button confirm-button">Continue Shopping</a>
-        </div>
-    </div>
-</div>
-
-</body>
-</html>
